requirements.txt
pygame==2.6.1
psycopg2-binary==2.9.11
PyQt5==5.15.9

